Reorganise components to adhear to progressive enhancement
