+++ templates/channel_manager/_index_right_col_content.html
@@ -7,7 +7,7 @@
     <p><a href="#"><img src="{% static 'images/social-icons/discord.svg' %}" class="social-icon mr-1" /> Veems on Discord</a></p>
     <p><a href="#"><img src="{% static 'images/social-icons/twitter.svg' %}" class="social-icon mr-1" /> Veems on Twitter</a></p>
 </div>
-<div class="border my-3 rounded p-3">
+<div id="actions" class="border my-3 rounded p-3">
     <h4>Actions</h4>
     <div><a href="#" class="btn btn-outline-secondary" data-toggle="modal" data-target="#uploadModal"><i
                 class="material-icons align-middle">backup</i> Upload
@@ -17,6 +17,12 @@
             Sync
             YouTube
             Videos</a></div>
-    <div><a href="#" data-toggle="modal" data-target="#createChannelModal" class="mt-2 btn btn-outline-secondary"><i class="material-icons align-middle">add_circle_outline</i> Create Channel</a></div>
+    <div id="create-channel-button"><a href="#" data-toggle="modal" data-target="#createChannelModal" class="mt-2 btn btn-outline-secondary"><i class="material-icons align-middle">add_circle_outline</i> Create Channel</a></div>
     <div><a href="{% url 'logout' %}" class="mt-2 btn btn-outline-secondary"><i class="material-icons align-middle">exit_to_app</i> Logout</a></div>
 </div>
+{{ block.super }}
+<script type="text/javascript">
+    app.actions({
+        element: document.getElementById('create-channel-button'),
+    })
+</script>
