+++ templates/channel_manager/index.html
@@ -98,11 +98,4 @@
     });
 </script>
 
-{{ block.super }}
-<script type="text/javascript">
-    app.ChannelDashboard({
-        element: document.getElementById('channel-container'),
-    })
-</script>
-
 {% endblock %}
